pyproject.toml
[project]
name = "caltrain-bot"
version = "0.1.0"
description = "Caltrain schedule bot for Telegram"
readme = "README.md"
authors = [
    { name = "Dima Timofeev", email = "1127655+CuriousDima@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings>=2.13.1",
    "pygtfs",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "sqlparse>=0.5.5",
]

[project.scripts]
caltrain-bot = "caltrain_bot:main"

[build-system]
requires = ["uv_build>=0.10.10,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.6",
    "ty>=0.0.23",
]

[tool.uv.sources]
pygtfs = { git = "https://github.com/CuriousDima/pygtfs", branch = "master" }
